ENH: Add "elxSupportedImageDimensions.h"
Made the sequence of supported dimensions `SupportedFixedImageDimensionSequence` accessible outside "elxTransformFactoryRegistration.cxx"
