+++ Common/Transforms/elxTransformFactoryRegistration.cxx
@@ -23,79 +23,13 @@
 #include <../Components/Transforms/EulerStackTransform/itkEulerStackTransform.h>
 #include <../Components/Transforms/TranslationStackTransform/itkTranslationStackTransform.h>
 
-#include <elxSupportedImageTypes.h>
+#include "elxSupportedImageDimensions.h"
 
 #include <itkBSplineTransform.h>
 #include <itkTransformFactory.h>
-#include <utility> // For index_sequence
 
 namespace
 {
-
-// The maximum possible supported image dimension (for either fixed or moving images).
-constexpr unsigned maxSupportedImageDimension{ 4 };
-
-
-template <unsigned VDimension, std::size_t... VIndex>
-constexpr bool
-SupportsFixedDimensionByImageTypeIndexSequence(std::index_sequence<VIndex...>)
-{
-  const bool foundEntries[] = { (elx::ElastixTypedef<VIndex + 1>::FixedDimension == VDimension)... };
-
-  for (const bool isDimensionFound : foundEntries)
-  {
-    if (isDimensionFound)
-    {
-      return true;
-    }
-  }
-  return false;
-}
-
-
-// Tells whether any of the supported fixed image types has the specified dimension.
-template <unsigned VDimension>
-constexpr bool
-SupportsFixedDimension()
-{
-  return SupportsFixedDimensionByImageTypeIndexSequence<VDimension>(
-    std::make_index_sequence<elx::NrOfSupportedImageTypes>());
-}
-
-
-template <unsigned VDimension = 2>
-struct FixedImageDimensionSupport
-{
-  // Adds those dimensions from the specified `dimensionSequence` that are supported.
-  template <std::size_t... VIndex>
-  constexpr static auto
-  AddSupportedDimensions(std::index_sequence<VIndex...> dimensionSequence)
-  {
-    using AddDimensionIfSupported = std::conditional_t<SupportsFixedDimension<VDimension>(),
-                                                       std::index_sequence<VDimension, VIndex...>,
-                                                       std::index_sequence<VIndex...>>;
-
-    return FixedImageDimensionSupport<VDimension + 1>::AddSupportedDimensions(AddDimensionIfSupported());
-  }
-};
-
-
-template <>
-struct FixedImageDimensionSupport<maxSupportedImageDimension + 1>
-{
-  template <std::size_t... VIndex>
-  constexpr static auto
-  AddSupportedDimensions(std::index_sequence<VIndex...> dimensionSequence)
-  {
-    return dimensionSequence;
-  }
-};
-
-
-/** A sequence of the dimensions of supported fixed images */
-const auto SupportedFixedImageDimensionSequence =
-  FixedImageDimensionSupport<>::AddSupportedDimensions(std::index_sequence<>());
-
 
 template <template <unsigned> class TTransform, std::size_t... VDimension>
 static void
@@ -114,7 +48,7 @@ template <template <unsigned> class TTransform>
 void
 RegisterTransform()
 {
-  RegisterTransformForEachDimension<TTransform>(SupportedFixedImageDimensionSequence);
+  RegisterTransformForEachDimension<TTransform>(elx::SupportedFixedImageDimensionSequence);
 }
 
 template <template <unsigned> class... TTransform>
